Make it so the addons with exclamation mark prefix aren't loaded

+++ SkEditor/API/AddonLoader.cs
@@ -31,6 +31,7 @@ public class AddonLoader
         IEnumerable<string> addonFolders = Directory.EnumerateDirectories(addonFolder, "*", SearchOption.TopDirectoryOnly);
         foreach (string folder in addonFolders)
         {
+            if (Path.GetFileName(folder).StartsWith('!')) continue;
             if (ApiVault.Get().GetAppConfig().AddonsToDelete.Contains(Path.GetFileName(folder)))
             {
                 Directory.Delete(folder, true);
@@ -64,9 +65,21 @@ public class AddonLoader
                 continue;
             }
 
-            Directory.Delete(folderWithoutPrefixPath, true);
-            ZipFile.ExtractToDirectory(updatedAddon, addonFolder);
-            File.Delete(updatedAddon);
+            try
+            {
+                Directory.Delete(folderWithoutPrefixPath, true);
+                ZipFile.ExtractToDirectory(updatedAddon, addonFolder);
+                File.Delete(updatedAddon);
+            }
+            catch (DirectoryNotFoundException)
+            {
+                Log.Warning($"Found \"{updatedAddon}\" in addons folder, but its folder \"{folderWithoutPrefixPath}\" doesn't exist. Deleting it.");
+                File.Delete(updatedAddon);
+            }
+            catch (Exception ex)
+            {
+                Log.Error(ex, $"Failed to update addon \"{folderWithoutPrefixPath}\"");
+            }
         }
     }
 
@@ -113,6 +126,7 @@ public class AddonLoader
         foreach (string dllFile in dllFiles)
         {
             string fileName = Path.GetFileName(dllFile);
+            if (fileName.StartsWith('!')) continue;
 
             if (ApiVault.Get().GetAppConfig().AddonsToDelete.Contains(fileName))
             {
